requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
Farama-Notifications==0.0.4
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.5.1
grpcio==1.73.0
gymnasium==1.1.1
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mazelib==0.9.16
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pettingzoo==1.25.0
pillow==11.2.1
protobuf==6.31.1
pydantic==2.11.5
pydantic_core==2.33.2
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
stable_baselines3==2.6.0
starlette==0.46.2
SuperSuit==3.9.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tinyscaler==1.2.8
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
